refactor(frontend): ConnectionConfig/SshConfig 型を types/connectionForm.ts に切り出し (#436)
UI コンポーネント ConnectionDialog.tsx から型 export していた `ConnectionConfig / SshConfig` を中立な `types/` 配下に移動。
`hook / reducer / 子コンポーネント` が UI コンポーネントの型 export に依存する状態を解消し、`hierarchical-architecture` のピラミッド構造に揃える。
