+++ frontend/src/components/layout/MainLayout.tsx
@@ -6,9 +6,9 @@ import { applyConnectionMigration } from '../../store/connectionMigration';
 import { useConnectionStore } from '../../store/connectionStore';
 import { useQueryStore } from '../../store/queryStore';
 import { useSessionStore } from '../../store/sessionStore';
+import type { ConnectionConfig } from '../../types/connectionForm';
 import { lazyWithRetry } from '../../utils/lazyWithRetry';
 import { checkQueryExecutability } from '../../utils/queryExecutionCheck';
-import type { ConnectionConfig } from '../dialogs/ConnectionDialog';
 import { QueryConfirmDialog } from '../dialogs/QueryConfirmDialog';
 import { ToolbarIcons } from '../icons/SvgIcons';
 import { CenterPanel } from './CenterPanel';
